removed prerender; introduced Renderer
