+++ scarlet/testing/deblend.py
@@ -47,7 +47,6 @@ def deblend(data: Dict[str, np.ndarray], max_iter: int, e_rel: float):
         edge_distance=None,
         min_snr=30,
         thresh=1,
-        prerender=False,
         fallback=True,
         silent=True,
         set_spectra=True,
@@ -62,19 +61,21 @@ def deblend(data: Dict[str, np.ndarray], max_iter: int, e_rel: float):
     if hasattr(observation, "log_norm"):
         log_norm = observation.log_norm
     else:
+        # TODO: not quite right, also bitrott since observation *has* log_norm
         _weights = observation.weights
         _images = observation.data
         log_sigma = np.zeros(_weights.shape, dtype=_weights.dtype)
         cuts = _weights > 0
         log_sigma[cuts] = np.log(1 / weights[cuts])
         log_norm = (
-            np.prod(_images.shape) / 2 * np.log(2 * np.pi) + np.sum(log_sigma) / 2
+            np.prod(_images[cuts].shape) / 2 * np.log(2 * np.pi) + np.sum(log_sigma) / 2
         )
 
     measurements = {
         "init time": (t1 - t0) * 1000,
         "runtime": (t2 - t1) * 1000 / len(sources),
         "iterations": len(blend.loss),
+        # log_norm is included in loss, keeping it here for backward compatibility of measurements
         "logL": blend.loss[-1] - log_norm,
         "init logL": blend.loss[0] - log_norm,
         # TODO: adding the number of skipped sources would be helpful
